Make GameUI layout resolution-independent
This is initial work towards scalable UI support. With this patch, it’s
possible to increase GAME_WIDTH and GAME_HEIGHT, and the user interface
largely gracefully adapts. However, menu code has not been updated, and
there are some other small graphical problems, so this capability is not
yet exposed anywhere.