Close the last remaining soundness hole in the tactical overlay
It is not entirely clear whether or not this is a good change, as it
makes the algorithm significantly more conservative. In situations where
the player is confident they will not be bombed (perhaps because the
opponent cannot have a stellar bomb available), it may feel
frustratingly so. But I am going to give it a try and see if it feels
okay or not in practice, and I can change it later if it proves
annoying.